requirements.txt
requests
openpyxl
psycopg2-binary
customtkinter
